Enable utm params
SXSW-94
